Reorganise roles code
