+++ cosmetics-web/app/models/search_user.rb
@@ -1,4 +1,6 @@
 class SearchUser < User
+  include SearchRolesConcern
+
   INVITATION_EXPIRATION_DAYS = 14
   ALLOW_INTERNATIONAL_PHONE_NUMBER = false
   TOTP_ISSUER = "Search Cosmetics".freeze
@@ -22,30 +24,6 @@ class SearchUser < User
   validates :mobile_number,
             phone: { message: :invalid, allow_international: ALLOW_INTERNATIONAL_PHONE_NUMBER },
             if: -> { mobile_number.present? }
-
-  def poison_centre_user?
-    poison_centre?
-  end
-
-  def msa_user?
-    msa?
-  end
-
-  def opss_science_user?
-    opss_science?
-  end
-
-  def can_view_product_ingredients?
-    !msa_user?
-  end
-
-  def can_view_nanomaterial_notification_files?
-    opss_science_user?
-  end
-
-  def can_view_nanomaterial_review_period_end_date?
-    msa_user? || opss_science_user?
-  end
 
   def resend_account_setup_link
     SearchNotifyMailer.invitation_email(self).deliver_later
